corrected rebond_dis_64 and added rebound to icon list
Rebound was always "half" in the large size filelist, with a dis but not a std. I added the std before but didn't add both to the list. I didn't remove the existing line just in case (but it can probably be replaced with placefiller now). Also weirdly the rebound_dis was incorrectly coloured, which has been corrected.